+++ app/client/cli/debug.go
@@ -19,11 +19,15 @@ import (
 
 // TECHDEBT: Lowercase variables / constants that do not need to be exported.
 const (
-	PromptResetToGenesis         string = "ResetToGenesis"
-	PromptPrintNodeState         string = "PrintNodeState"
-	PromptTriggerNextView        string = "TriggerNextView"
-	PromptTogglePacemakerMode    string = "TogglePacemakerMode"
+	PromptResetToGenesis      string = "ResetToGenesis"
+	PromptPrintNodeState      string = "PrintNodeState"
+	PromptTriggerNextView     string = "TriggerNextView"
+	PromptTogglePacemakerMode string = "TogglePacemakerMode"
+
 	PromptShowLatestBlockInStore string = "ShowLatestBlockInStore"
+
+	PromptSendMetadataRequest string = "MetadataRequest"
+	PromptSendBlockRequest    string = "BlockRequest"
 )
 
 var (
@@ -36,6 +40,8 @@ var (
 		PromptTriggerNextView,
 		PromptTogglePacemakerMode,
 		PromptShowLatestBlockInStore,
+		PromptSendMetadataRequest,
+		PromptSendBlockRequest,
 	}
 
 	// validators holds the list of the validators at genesis time so that we can use it to create a debug address book provider.
@@ -159,6 +165,18 @@ func handleSelect(selection string) {
 			Message: nil,
 		}
 		sendDebugMessage(m)
+	case PromptSendMetadataRequest:
+		m := &messaging.DebugMessage{
+			Action:  messaging.DebugMessageAction_DEBUG_CONSENSUS_SEND_METADATA_REQ,
+			Message: nil,
+		}
+		broadcastDebugMessage(m)
+	case PromptSendBlockRequest:
+		m := &messaging.DebugMessage{
+			Action:  messaging.DebugMessageAction_DEBUG_CONSENSUS_SEND_BLOCK_REQ,
+			Message: nil,
+		}
+		broadcastDebugMessage(m)
 	default:
 		logger.Global.Error().Msg("Selection not yet implemented...")
 	}