+++ persistence/docs/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.0.0.33] - 2023-02-09
+
+- Added mock generation to the `kvstore/kvstore.go`.
+
 ## [0.0.0.32] - 2023-02-07
 
 - Minor documentation cleanup